+++ Controller/ConnectController.php
@@ -14,25 +14,25 @@ namespace HWI\Bundle\OAuthBundle\Controller;
 use HWI\Bundle\OAuthBundle\OAuth\ResourceOwnerInterface;
 use HWI\Bundle\OAuthBundle\Security\Core\Authentication\Token\OAuthToken;
 use HWI\Bundle\OAuthBundle\Security\Core\Exception\AccountNotLinkedException;
-use Symfony\Component\DependencyInjection\ContainerAware;
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Exception\AccountStatusException;
-use Symfony\Component\Security\Core\SecurityContext;
+use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\SecurityContextInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Http\Event\InteractiveLoginEvent;
 use Symfony\Component\Security\Http\SecurityEvents;
 
 /**
- * ConnectController
- *
  * @author Alexander <iam.asm89@gmail.com>
  */
-class ConnectController extends ContainerAware
+class ConnectController extends Controller
 {
     /**
      * Action that handles the login 'form'. If connecting is enabled the
@@ -45,7 +45,7 @@ class ConnectController extends ContainerAware
     public function connectAction(Request $request)
     {
         $connect = $this->container->getParameter('hwi_oauth.connect');
-        $hasUser = $this->container->get('security.context')->isGranted('IS_AUTHENTICATED_REMEMBERED');
+        $hasUser = $this->isGranted('IS_AUTHENTICATED_REMEMBERED');
 
         $error = $this->getErrorForRequest($request);
 
@@ -58,7 +58,7 @@ class ConnectController extends ContainerAware
             $session = $request->getSession();
             $session->set('_hwi_oauth.registration_error.'.$key, $error);
 
-            return new RedirectResponse($this->generate('hwi_oauth_connect_registration', array('key' => $key)));
+            return $this->redirectToRoute('hwi_oauth_connect_registration', array('key' => $key));
         }
 
         if ($error) {
@@ -66,7 +66,7 @@ class ConnectController extends ContainerAware
             $error = $error->getMessage();
         }
 
-        return $this->container->get('templating')->renderResponse('HWIOAuthBundle:Connect:login.html.' . $this->getTemplatingEngine(), array(
+        return $this->render('HWIOAuthBundle:Connect:login.html.'.$this->getTemplatingEngine(), array(
             'error'   => $error,
         ));
     }
@@ -91,7 +91,7 @@ class ConnectController extends ContainerAware
             throw new NotFoundHttpException();
         }
 
-        $hasUser = $this->container->get('security.context')->isGranted('IS_AUTHENTICATED_REMEMBERED');
+        $hasUser = $this->isGranted('IS_AUTHENTICATED_REMEMBERED');
         if ($hasUser) {
             throw new AccessDeniedException('Cannot connect already registered account.');
         }
@@ -100,7 +100,7 @@ class ConnectController extends ContainerAware
         $error = $session->get('_hwi_oauth.registration_error.'.$key);
         $session->remove('_hwi_oauth.registration_error.'.$key);
 
-        if (!($error instanceof AccountNotLinkedException) || (time() - $key > 300)) {
+        if (!$error instanceof AccountNotLinkedException || time() - $key > 300) {
             // todo: fix this
             throw new \Exception('Cannot register an account.');
         }
@@ -124,7 +124,7 @@ class ConnectController extends ContainerAware
             // Authenticate the user
             $this->authenticateUser($request, $form->getData(), $error->getResourceOwnerName(), $error->getRawToken());
 
-            return $this->container->get('templating')->renderResponse('HWIOAuthBundle:Connect:registration_success.html.' . $this->getTemplatingEngine(), array(
+            return $this->render('HWIOAuthBundle:Connect:registration_success.html.'.$this->getTemplatingEngine(), array(
                 'userInformation' => $userInformation,
             ));
         }
@@ -133,7 +133,7 @@ class ConnectController extends ContainerAware
         $key = time();
         $session->set('_hwi_oauth.registration_error.'.$key, $error);
 
-        return $this->container->get('templating')->renderResponse('HWIOAuthBundle:Connect:registration.html.' . $this->getTemplatingEngine(), array(
+        return $this->render('HWIOAuthBundle:Connect:registration.html.'.$this->getTemplatingEngine(), array(
             'key' => $key,
             'form' => $form->createView(),
             'userInformation' => $userInformation,
@@ -160,7 +160,7 @@ class ConnectController extends ContainerAware
             throw new NotFoundHttpException();
         }
 
-        $hasUser = $this->container->get('security.context')->isGranted('IS_AUTHENTICATED_REMEMBERED');
+        $hasUser = $this->isGranted('IS_AUTHENTICATED_REMEMBERED');
         if (!$hasUser) {
             throw new AccessDeniedException('Cannot connect an account.');
         }
@@ -183,6 +183,11 @@ class ConnectController extends ContainerAware
             $accessToken = $session->get('_hwi_oauth.connect_confirmation.'.$key);
         }
 
+        // Redirect to the login path if the token is empty (Eg. User cancelled auth)
+        if (null === $accessToken) {
+            return $this->redirectToRoute($this->container->getParameter('hwi_oauth.failed_auth_path'));
+        }
+        
         $userInformation = $resourceOwner->getUserInformation($accessToken);
 
         // Show confirmation page?
@@ -192,37 +197,34 @@ class ConnectController extends ContainerAware
 
         // Handle the form
         /** @var $form FormInterface */
-        $form = $this->container->get('form.factory')
-            ->createBuilder('form')
-            ->getForm();
+        $form = $this->createForm('form');
 
-        if ($request->isMethod('POST')) {
-            $form->bind($request);
+        $form->handleRequest($request);
 
-            if ($form->isValid()) {
-                show_confirmation_page:
+        if ($form->isSubmitted() && $form->isValid()) {
+            show_confirmation_page:
 
-                /** @var $currentToken OAuthToken */
-                $currentToken = $this->container->get('security.context')->getToken();
-                $currentUser  = $currentToken->getUser();
+            /** @var $currentToken OAuthToken */
+            $currentToken = $this->getToken();
+            $currentUser = $currentToken->getUser();
 
-                $this->container->get('hwi_oauth.account.connector')->connect($currentUser, $userInformation);
+            $this->container->get('hwi_oauth.account.connector')->connect($currentUser, $userInformation);
 
-                if ($currentToken instanceof OAuthToken) {
-                    // Update user token with new details
-                    $this->authenticateUser($request, $currentUser, $service, $currentToken->getRawToken(), false);
-                }
-
-                return $this->container->get('templating')->renderResponse('HWIOAuthBundle:Connect:connect_success.html.' . $this->getTemplatingEngine(), array(
-                    'userInformation' => $userInformation,
-                ));
+            if ($currentToken instanceof OAuthToken) {
+                // Update user token with new details
+                $this->authenticateUser($request, $currentUser, $service, $currentToken->getRawToken(), false);
             }
+
+            return $this->render('HWIOAuthBundle:Connect:connect_success.html.' . $this->getTemplatingEngine(), array(
+                'userInformation' => $userInformation,
+                'service' => $service,
+            ));
         }
 
-        return $this->container->get('templating')->renderResponse('HWIOAuthBundle:Connect:connect_confirm.html.' . $this->getTemplatingEngine(), array(
-            'key'             => $key,
-            'service'         => $service,
-            'form'            => $form->createView(),
+        return $this->render('HWIOAuthBundle:Connect:connect_confirm.html.' . $this->getTemplatingEngine(), array(
+            'key' => $key,
+            'service' => $service,
+            'form' => $form->createView(),
             'userInformation' => $userInformation,
         ));
     }
@@ -243,20 +245,21 @@ class ConnectController extends ContainerAware
             $session = $request->getSession();
             $session->start();
 
-            $providerKey = $this->container->getParameter('hwi_oauth.firewall_name');
-            $sessionKey = '_security.' . $providerKey . '.target_path';
+            foreach ($this->container->getParameter('hwi_oauth.firewall_names') as $providerKey) {
+                $sessionKey = '_security.'.$providerKey.'.target_path';
 
-            $param = $this->container->getParameter('hwi_oauth.target_path_parameter');
-            if (!empty($param) && $targetUrl = $request->get($param, null, true)) {
-                $session->set($sessionKey, $targetUrl);
-            }
+                $param = $this->container->getParameter('hwi_oauth.target_path_parameter');
+                if (!empty($param) && $targetUrl = $request->get($param)) {
+                    $session->set($sessionKey, $targetUrl);
+                }
 
-            if ($this->container->getParameter('hwi_oauth.use_referer') && !$session->has($sessionKey) && ($targetUrl = $request->headers->get('Referer')) && $targetUrl !== $authorizationUrl) {
-                $session->set($sessionKey, $targetUrl);
+                if ($this->container->getParameter('hwi_oauth.use_referer') && !$session->has($sessionKey) && ($targetUrl = $request->headers->get('Referer')) && $targetUrl !== $authorizationUrl) {
+                    $session->set($sessionKey, $targetUrl);
+                }
             }
         }
 
-        return new RedirectResponse($authorizationUrl);
+        return $this->redirect($authorizationUrl);
     }
 
     /**
@@ -268,12 +271,16 @@ class ConnectController extends ContainerAware
      */
     protected function getErrorForRequest(Request $request)
     {
+        // Symfony <2.6 BC. To be removed.
+        $authenticationErrorKey = class_exists('Symfony\Component\Security\Core\Security')
+            ? Security::AUTHENTICATION_ERROR : SecurityContextInterface::AUTHENTICATION_ERROR;
+
         $session = $request->getSession();
-        if ($request->attributes->has(SecurityContext::AUTHENTICATION_ERROR)) {
-            $error = $request->attributes->get(SecurityContext::AUTHENTICATION_ERROR);
-        } elseif (null !== $session && $session->has(SecurityContext::AUTHENTICATION_ERROR)) {
-            $error = $session->get(SecurityContext::AUTHENTICATION_ERROR);
-            $session->remove(SecurityContext::AUTHENTICATION_ERROR);
+        if ($request->attributes->has($authenticationErrorKey)) {
+            $error = $request->attributes->get($authenticationErrorKey);
+        } elseif (null !== $session && $session->has($authenticationErrorKey)) {
+            $error = $session->get($authenticationErrorKey);
+            $session->remove($authenticationErrorKey);
         } else {
             $error = '';
         }
@@ -292,17 +299,25 @@ class ConnectController extends ContainerAware
      */
     protected function getResourceOwnerByName($name)
     {
-        $ownerMap = $this->container->get('hwi_oauth.resource_ownermap.'.$this->container->getParameter('hwi_oauth.firewall_name'));
+        foreach ($this->container->getParameter('hwi_oauth.firewall_names') as $firewall) {
+            $id = 'hwi_oauth.resource_ownermap.'.$firewall;
+            if (!$this->container->has($id)) {
+                continue;
+            }
 
-        if (null === $resourceOwner = $ownerMap->getResourceOwnerByName($name)) {
-            throw new \RuntimeException(sprintf("No resource owner with name '%s'.", $name));
+            $ownerMap = $this->container->get($id);
+            if ($resourceOwner = $ownerMap->getResourceOwnerByName($name)) {
+                return $resourceOwner;
+            }
         }
 
-        return $resourceOwner;
+        throw new \RuntimeException(sprintf("No resource owner with name '%s'.", $name));
     }
 
     /**
      * Generates a route.
+     *
+     * @deprecated since version 0.4. Will be removed in 1.0.
      *
      * @param string  $route    Route name
      * @param array   $params   Route parameters
@@ -312,17 +327,19 @@ class ConnectController extends ContainerAware
      */
     protected function generate($route, $params = array(), $absolute = false)
     {
+        @trigger_error('The '.__METHOD__.' method is deprecated since version 0.4 and will be removed in 1.0. Use Symfony\Bundle\FrameworkBundle\Controller\Controller::generateUrl instead.', E_USER_DEPRECATED);
+
         return $this->container->get('router')->generate($route, $params, $absolute);
     }
 
     /**
-     * Authenticate a user with Symfony Security
+     * Authenticate a user with Symfony Security.
      *
      * @param Request       $request
      * @param UserInterface $user
      * @param string        $resourceOwnerName
      * @param string        $accessToken
-     * @param boolean       $fakeLogin
+     * @param bool          $fakeLogin
      */
     protected function authenticateUser(Request $request, UserInterface $user, $resourceOwnerName, $accessToken, $fakeLogin = true)
     {
@@ -338,7 +355,7 @@ class ConnectController extends ContainerAware
         $token->setUser($user);
         $token->setAuthenticated(true);
 
-        $this->container->get('security.context')->setToken($token);
+        $this->setToken($token);
 
         if ($fakeLogin) {
             // Since we're "faking" normal login, we need to throw our INTERACTIVE_LOGIN event manually
@@ -357,5 +374,61 @@ class ConnectController extends ContainerAware
     protected function getTemplatingEngine()
     {
         return $this->container->getParameter('hwi_oauth.templating.engine');
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * Symfony <2.6 BC. To be removed.
+     */
+    protected function isGranted($attributes, $object = null)
+    {
+        if (method_exists('Symfony\Bundle\FrameworkBundle\Controller\Controller', 'isGranted')) {
+            return parent::isGranted($attributes, $object);
+        }
+
+        return $this->get('security.context')->isGranted($attributes, $object);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * Symfony <2.6 BC. To be removed.
+     */
+    protected function redirectToRoute($route, array $parameters = array(), $status = 302)
+    {
+        if (method_exists('Symfony\Bundle\FrameworkBundle\Controller\Controller', 'redirectToRoute')) {
+            return parent::redirectToRoute($route, $parameters, $status);
+        }
+
+        return $this->redirect($this->generateUrl($route, $parameters), $status);
+    }
+
+    /**
+     * @return null|TokenInterface
+     *
+     * Symfony <2.6 BC. Remove it and use only security.token_storage service instead.
+     */
+    protected function getToken()
+    {
+        if ($this->has('security.token_storage')) {
+            return $this->get('security.token_storage')->getToken();
+        }
+
+        return $this->get('security.context')->getToken();
+    }
+
+    /**
+     * @param TokenInterface $token
+     *
+     * Symfony <2.6 BC. Remove it and use only security.token_storage service instead.
+     */
+    protected function setToken(TokenInterface $token)
+    {
+        if ($this->has('security.token_storage')) {
+            return $this->get('security.token_storage')->setToken($token);
+        }
+
+        return $this->get('security.context')->setToken($token);
     }
 }