+++ OAuth/ResourceOwner/FacebookResourceOwner.php
@@ -28,6 +28,8 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
     protected $paths = array(
         'identifier' => 'id',
         'nickname'   => 'name',
+        'firstname'   => 'first_name',
+        'lastname'   => 'last_name',
         'realname'   => 'name',
         'email'      => 'email',
     );
@@ -35,9 +37,30 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
     /**
      * {@inheritDoc}
      */
+    public function getUserInformation(array $accessToken, array $extraParameters = array())
+    {
+        if ($this->options['appsecret_proof']) {
+            $extraParameters['appsecret_proof'] = hash_hmac('sha256', $accessToken['access_token'], $this->options['client_secret']);
+        }
+
+        return parent::getUserInformation($accessToken, $extraParameters);
+    }
+    
+    /**
+     * {@inheritDoc}
+     */
     public function getAuthorizationUrl($redirectUri, array $extraParameters = array())
     {
-        return parent::getAuthorizationUrl($redirectUri, array_merge(array('display' => $this->options['display']), $extraParameters));
+        $extraOptions = array();
+        if (isset($this->options['display'])) {
+            $extraOptions['display'] = $this->options['display'];
+        }
+
+        if (isset($this->options['auth_type'])) {
+            $extraOptions['auth_type'] = $this->options['auth_type'];
+        }
+
+        return parent::getAuthorizationUrl($redirectUri, array_merge($extraOptions, $extraParameters));
     }
 
     /**
@@ -67,10 +90,9 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
             'client_secret' => $this->options['client_secret'],
         );
 
-        $response = $this->httpRequest($this->normalizeUrl($this->options['revoke_token_url'], array('token' => $token)), $parameters, array(), HttpRequestInterface::METHOD_POST);
-        $response = $this->getResponseContent($response);
+        $response = $this->httpRequest($this->normalizeUrl($this->options['revoke_token_url'], array('access_token' => $token)), $parameters, array(), HttpRequestInterface::METHOD_DELETE);
 
-        return 'true' == $response;
+        return 200 === $response->getStatusCode();
     }
 
     /**
@@ -81,19 +103,31 @@ class FacebookResourceOwner extends GenericOAuth2ResourceOwner
         parent::configureOptions($resolver);
 
         $resolver->setDefaults(array(
-            'authorization_url'   => 'https://www.facebook.com/dialog/oauth',
-            'access_token_url'    => 'https://graph.facebook.com/oauth/access_token',
-            'revoke_token_url'    => 'https://graph.facebook.com/me/permissions',
-            'infos_url'           => 'https://graph.facebook.com/me',
+            'authorization_url'   => 'https://www.facebook.com/v2.0/dialog/oauth',
+            'access_token_url'    => 'https://graph.facebook.com/v2.0/oauth/access_token',
+            'revoke_token_url'    => 'https://graph.facebook.com/v2.0/me/permissions',
+            'infos_url'           => 'https://graph.facebook.com/v2.0/me',
 
             'use_commas_in_scope' => true,
 
             'display'             => null,
+            'auth_type'           => null,
+            'appsecret_proof'     => false,
         ));
 
-        $resolver->setAllowedValues(array(
-            // @link https://developers.facebook.com/docs/reference/dialogs/#display
-            'display' => array('page', 'popup', 'touch', null),
-        ));
+        // Symfony <2.6 BC
+        if (method_exists($resolver, 'setDefined')) {
+            $resolver
+                ->setAllowedValues('display', array('page', 'popup', 'touch', null)) // @link https://developers.facebook.com/docs/reference/dialogs/#display
+                ->setAllowedValues('auth_type', array('rerequest', null)) // @link https://developers.facebook.com/docs/reference/javascript/FB.login/
+                ->setAllowedTypes('appsecret_proof', 'bool') // @link https://developers.facebook.com/docs/graph-api/securing-requests
+            ;
+        } else {
+            $resolver->setAllowedValues(array(
+                'display'   => array('page', 'popup', 'touch', null),
+                'auth_type' => array('rerequest', null),
+                'appsecret_proof' => array(true, false),
+            ));
+        }
     }
 }