+++ Tests/OAuth/ResourceOwner/GenericOAuth1ResourceOwnerTest.php
@@ -51,19 +51,34 @@ class GenericOAuth1ResourceOwnerTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
-    public function testGetInvalidOptionThrowsException()
+    public function testUndefinedOptionThrowsException()
     {
         $this->createResourceOwner($this->resourceOwnerName, array('non_existing' => null));
     }
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
     public function testInvalidOptionValueThrowsException()
     {
         $this->createResourceOwner($this->resourceOwnerName, array('csrf' => 'invalid'));
+    }
+
+    public function testHandleRequest()
+    {
+        $request = new Request(array('test' => 'test'));
+
+        $this->assertFalse($this->resourceOwner->handles($request));
+
+        $request = new Request(array('oauth_token' => 'test'));
+
+        $this->assertTrue($this->resourceOwner->handles($request));
+
+        $request = new Request(array('oauth_token' => 'test', 'test' => 'test'));
+
+        $this->assertTrue($this->resourceOwner->handles($request));
     }
 
     public function testGetUserInformation()
@@ -226,6 +241,26 @@ class GenericOAuth1ResourceOwnerTest extends \PHPUnit_Framework_TestCase
         $this->storage->expects($this->once())
             ->method('fetch')
             ->will($this->returnValue(array('oauth_token' => 'token', 'oauth_token_secret' => 'secret')));
+
+        $this->storage->expects($this->never())
+            ->method('save');
+
+        $request = new Request(array('oauth_token' => 'token', 'oauth_verifier' => 'code'));
+
+        $this->resourceOwner->getAccessToken($request, 'http://redirect.to/');
+    }
+
+    /**
+     * @expectedException \Symfony\Component\Security\Core\Exception\AuthenticationException
+     */
+    public function testGetAccessTokenInvalidArgumentException()
+    {
+        $this->storage->expects($this->once())
+            ->method('fetch')
+            ->will($this->throwException(new \InvalidArgumentException));
+
+        $this->buzzClient->expects($this->never())
+            ->method('send');
 
         $this->storage->expects($this->never())
             ->method('save');