+++ Tests/OAuth/ResourceOwner/GoogleResourceOwnerTest.php
@@ -26,6 +26,8 @@ json;
         'identifier'     => 'id',
         'nickname'       => 'name',
         'realname'       => 'name',
+        'firstname'      => 'given_name',
+        'lastname'       => 'family_name',
         'email'          => 'email',
         'profilepicture' => 'picture',
     );
@@ -36,7 +38,7 @@ json;
     );
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
     public function testInvalidAccessTypeOptionValueThrowsException()
     {
@@ -44,7 +46,7 @@ json;
     }
 
     /**
-     * @expectedException \Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     * @expectedException \Symfony\Component\OptionsResolver\Exception\ExceptionInterface
      */
     public function testInvalidApprovalPromptOptionValueThrowsException()
     {